requirements.txt
streamlit>=1.35.0
requests>=2.31.0
pyyaml>=6.0
python-dateutil>=2.8.2
pandas>=2.0.0
python-dotenv>=1.0.0
supabase>=2.0.0
